Covert for 2021 Census
Download targets now use 2021 census information rather than 2016.

GeoPandas changed some behaviour which was updated:
  * op= is now predicate= for spatial joins
  * GeoDataFrame.append is now pd.concat([*geodataframe])

I also made some changes to process just Manitoba and Saskatchwan due to
local resource constraints pending running the repo on a more powerful
machine.

I started the process of using make to invoke the process, as I want to
depend on this repo as a submodule of another which depends on make and
it greatly facilitates the integration. This is not a finished process
though.