Refactor code to loops, not working
Now the massive blocks of repeated code are gone, in prep for
moving over to a Makefile driven analysis.

This "works" in that it builds, but it doesn't work in that there
are several serious errors and I'm not really sure what is wrong.

These will need to be quashed before I can really trust it.