Get correspondances and remove downloaded files
I think I will need these correspondance files to update the ODB
buildings. This isn't done yet as it's not required for Manitoba which
is my current focus. I think you will need to match the 2016 values to
the census you are currently using, which is 2021 right now.

I also added a cleaning target to the Makefile for deleting the
downloaded information because otherwise it's a pain to test.